rag-experiment/requirements.txt

torch
transformers
faiss-cpu
numpy
